[backend] Refactor file structure
